fixing curl tests
