+++ app/client/cypress/support/Pages/DataSources.ts
@@ -37,6 +37,7 @@ export enum Widgets {
   Text,
   WDSTable,
 }
+
 type AppModes = "Edit" | "View";
 
 interface RunQueryParams {
@@ -44,6 +45,7 @@ interface RunQueryParams {
   expectedStatus?: boolean;
   waitTimeInterval?: number;
 }
+
 export class DataSources {
   private agHelper = ObjectsRegistry.AggregateHelper;
   private table = ObjectsRegistry.Table;
@@ -167,7 +169,6 @@ export class DataSources {
     dbName +
     "']/ancestor::div[contains(@class, 't--mock-datasource')][1]";
   private _createBlankGraphQL = ".t--createBlankApiGraphqlCard";
-  private _createBlankCurl = ".t--createBlankCurlCard";
   private _graphQLHeaderKey = "input[name='headers[0].key']";
   private _graphQLHeaderValue = "input[name='headers[0].value']";
   _graphqlQueryEditor = ".t--graphql-query-editor";
@@ -212,8 +213,6 @@ export class DataSources {
     widgetType +
     "')]";
 
-  private _curlTextArea =
-    "//label[text()='Paste CURL Code Here']/parent::form/div";
   _noSchemaAvailable = (dbName: string) =>
     "//div[text()='" +
     dbName +
@@ -681,19 +680,6 @@ export class DataSources {
     );
   }
 
-  public FillCurlNImport(value: string) {
-    this.NavigateToDSCreateNew();
-    this.agHelper.GetNClick(this._createBlankCurl);
-    this.ImportCurlNRun(value);
-  }
-
-  public ImportCurlNRun(value: string) {
-    this.agHelper.UpdateTextArea(this._curlTextArea, value);
-    this.agHelper.Sleep(500); //Clicking import after value settled
-    this.agHelper.ClickButton("Import");
-    this.apiPage.RunAPI();
-  }
-
   public FillFirestoreDSForm(environment = this.dataManager.defaultEnviorment) {
     this.agHelper.TypeText(
       this.locator._inputFieldByName("Database URL") +
@@ -1015,6 +1001,7 @@ export class DataSources {
     this.assertHelper.AssertNetworkStatus("@getConsolidatedData", 200);
     this.assertHelper.AssertNetworkStatus("getWorkspace");
   }
+
   public ReconnectModalValidation(
     dbName: string,
     dsName: "PostgreSQL" | "MySQL" | "MongoDB",
